requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.3
